feat: create search feature to use for feed and chat
