--- lib/app/features/feed/feed_search/views/pages/feed_simple_search_page/components/feed_search_history/feed_search_history_query_list_item.dart
@@ -1,42 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import 'package:flutter/widgets.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
-import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/router/app_routes.dart';
-import 'package:ion/generated/assets.gen.dart';
-
-class FeedSearchHistoryQueryListItem extends ConsumerWidget {
-  const FeedSearchHistoryQueryListItem({required this.query, super.key});
-
-  final String query;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return GestureDetector(
-      behavior: HitTestBehavior.opaque,
-      onTap: () {
-        FeedSimpleSearchRoute(query: query).replace(context);
-      },
-      child: Padding(
-        padding: EdgeInsets.symmetric(
-          vertical: 8.0.s,
-          horizontal: ScreenSideOffset.defaultSmallMargin,
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Text(
-              query,
-              style: context.theme.appTextThemes.body2.copyWith(
-                color: context.theme.appColors.primaryText,
-              ),
-            ),
-            Assets.svg.iconSearchHistorylink.icon(size: 20.0.s),
-          ],
-        ),
-      ),
-    );
-  }
-}
