--- lib/app/features/feed/feed_search/views/pages/feed_simple_search_page/components/feed_search_history_empty/feed_search_history_empty.dart
@@ -1,31 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import 'package:flutter/widgets.dart';
-import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
-import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/generated/assets.gen.dart';
-
-class FeedSearchHistoryEmpty extends StatelessWidget {
-  const FeedSearchHistoryEmpty({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: ScreenSideOffset.small(
-        child: Column(
-          children: [
-            SizedBox(height: 150.0.s),
-            Assets.svg.walletIconWalletSearching.icon(size: 48.0.s),
-            SizedBox(height: 8.0.s),
-            Text(
-              context.i18n.feed_search_empty,
-              style: context.theme.appTextThemes.body2.copyWith(
-                color: context.theme.appColors.tertararyText,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
