+++ lib/app/features/feed_search/model/feed_advanced_search_category.dart
