+++ lib/app/router/app_routes.dart
@@ -37,10 +37,6 @@ import 'package:ion/app/features/feed/create_story/create_story_modal.dart';
 import 'package:ion/app/features/feed/create_story/views/pages/story_camera_widget.dart';
 import 'package:ion/app/features/feed/create_story/views/pages/story_preview_page.dart';
 import 'package:ion/app/features/feed/create_video/views/pages/create_video_modal/create_video_modal.dart';
-import 'package:ion/app/features/feed/feed_search/views/pages/feed_advanced_search_page/feed_advanced_search_page.dart';
-import 'package:ion/app/features/feed/feed_search/views/pages/feed_search_filters_page/feed_search_filters_page.dart';
-import 'package:ion/app/features/feed/feed_search/views/pages/feed_search_languages_page/feed_search_languages_page.dart';
-import 'package:ion/app/features/feed/feed_search/views/pages/feed_simple_search_page/feed_simple_search_page.dart';
 import 'package:ion/app/features/feed/views/pages/comment_post_modal/comment_post_modal.dart';
 import 'package:ion/app/features/feed/views/pages/feed_main_modal/feed_main_modal_page.dart';
 import 'package:ion/app/features/feed/views/pages/feed_page/feed_page.dart';
@@ -48,6 +44,11 @@ import 'package:ion/app/features/feed/views/pages/post_details_page/post_details
 import 'package:ion/app/features/feed/views/pages/post_reply_modal/post_reply_modal.dart';
 import 'package:ion/app/features/feed/views/pages/repost_options_modal/repost_options_modal.dart';
 import 'package:ion/app/features/feed/views/pages/share_post_modal/share_post_modal.dart';
+import 'package:ion/app/features/feed_search/views/pages/chat_simple_search_page/chat_simple_search_page.dart';
+import 'package:ion/app/features/feed_search/views/pages/feed_advanced_search_page/feed_advanced_search_page.dart';
+import 'package:ion/app/features/feed_search/views/pages/feed_search_filters_page/feed_search_filters_page.dart';
+import 'package:ion/app/features/feed_search/views/pages/feed_search_languages_page/feed_search_languages_page.dart';
+import 'package:ion/app/features/feed_search/views/pages/feed_simple_search_page/feed_simple_search_page.dart';
 import 'package:ion/app/features/gallery/views/pages/media_picker_page.dart';
 import 'package:ion/app/features/protect_account/authenticator/data/model/authenticator_steps.dart';
 import 'package:ion/app/features/protect_account/authenticator/views/pages/delete_authenticator/authenticator_delete_page.dart';
